fix(nc-gui): move fields and attachmentCol to useLTARStore
