+++ packages/nc-gui/components/virtual-cell/components/UnLinkedItems.vue
@@ -45,7 +45,8 @@ const {
   row,
   resetChildrenExcludedOffsetCount,
   loadRelatedTableMeta,
-  targetViewColumnsById,
+  attachmentCol,
+  fields,
 } = useLTARStoreOrThrow()
 
 const { addLTARRef, isNew, removeLTARRef, state: rowState } = useSmartsheetRowStoreOrThrow()
@@ -164,23 +165,6 @@ const newRowState = computed(() => {
       [colInRelatedTable.title as string]: row?.value && [row.value.row],
     }
   }
-})
-
-const attachmentCol = computedInject(FieldsInj, (_fields) => {
-  return (relatedTableMeta.value.columns ?? []).filter((col) => isAttachment(col))[0]
-})
-
-const fields = computedInject(FieldsInj, (_fields) => {
-  return (relatedTableMeta.value.columns ?? [])
-    .filter((col) => !isSystemColumn(col) && !isPrimary(col) && !isLinksOrLTAR(col) && !isAttachment(col))
-    .sort((a, b) => {
-      if (isPublic.value) {
-        return (a.meta?.defaultViewColOrder ?? Infinity) - (b.meta?.defaultViewColOrder ?? Infinity)
-      }
-
-      return (targetViewColumnsById.value[a.id!]?.order ?? Infinity) - (targetViewColumnsById.value[b.id!]?.order ?? Infinity)
-    })
-    .slice(0, isMobileMode.value ? 1 : 3)
 })
 
 const totalItemsToShow = computed(() => {
